Fixed inappropriate naming.
